Potential fix for code scanning alert no. 18: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using a larger type before the result is cast to size_t. This can be achieved by casting one of the operands to size_t before performing the multiplication. This way, the multiplication will be done in the size_t type, which is typically larger than unsigned int and can hold larger values without overflowing.

+++ libopenjpeg-2.3.1/src/lib/openjp2/j2k.c
@@ -7151,7 +7151,7 @@ OPJ_BOOL opj_j2k_setup_encoder(opj_j2k_t *p_j2k,
 
     /* initialize the mutiple tiles */
     /* ---------------------------- */
-    cp->tcps = (opj_tcp_t*) opj_calloc(cp->tw * cp->th, sizeof(opj_tcp_t));
+    cp->tcps = (opj_tcp_t*) opj_calloc((size_t)cp->tw * cp->th, sizeof(opj_tcp_t));
     if (!cp->tcps) {
         opj_event_msg(p_manager, EVT_ERROR,
                       "Not enough memory to allocate tile coding parameters\n");
